Cosmetic changes

+++ TaskManager/Date.cpp
@@ -4,8 +4,8 @@ using namespace std;
 
 Date::Date()
 {
-    day = (uint8_t)0;
-    month = (uint8_t)0;
+    day = 0;
+    month = 0;
     year = 0;
 }
 Date::Date(uint8_t day, uint8_t month, int year)
